// src/@types/embed.d.ts
export interface ctaStyle {
  buttonText: string;
  alignment: string;
  textStyle: string;
  textColor: string;
  buttonColor: string;
  buttonTextColor: string;
  borderRadius: string;
  ctaTop: number;
  ctaLeft: number;
}

export interface buttonStyle {
  backgroundColor: string;
  textColor: string;
  borderRadius: string;
  btnTop: string;
  btnLeft: string;
}

export interface titleStyle {
  estiloHL: string;
  corHL: string;
}

export interface subTitleStyle {
  alinhamentoSHL: string;
  estiloSHL: string;
  corSHL: string;
}

export interface ctaPStyle {
  alinhamentoCta: string;
  estiloCta: string;
  styleCta: string;
  corCta: string;
  delay: number;
}

export interface ctaBtnStyle {
  corBG: string;
  corTX: string;
  borderRadius: number;
}